requirements.txt
oracledb>=1.3.0
PyYAML
tqdm 